Added batch parameter through screens.
It's nice to have a sort of global batch
that is passed around through the project so that
at the startup of the game textures can be loaded one time
and can be used for the whole game session.

There is a big problem though in the implementation of
loading of textures in rungame, because textures seem to be loaded
at every frame, which causes a huge memory leak.

A better approach would be to load them at the start of the game
in the gBatch and then pass them around during the session.